Add User Interface
This commit adds UI to the Rock Paper Scissors
game. The goal was to practice using the DOM to
manipulate the page and to practice using event
listeners to respond to user input.

The game is now fully functional and can be played
by clicking on the buttons on the page.

The following changes were made:
- Updated instruction text for new mode of play
- Added dynamic trackers for:
    - Round Number
    - Player & Computer Choices
    - Player & Computer Score
    - Round Winner
- Added text for the final game winner
- Added Restart button to reset the game